refactor: Left, Right 모달 병합 & ButtonCategoryCommon 수정
