+++ src/shared/components/AddCategoryListModal.tsx
@@ -1,9 +1,20 @@
 import { RefObject, useState } from 'react';
 
-import CategoryModalLeft from '@/shared/components/CategoryModalLeft';
-import CategoryModalRight from '@/shared/components/CategoryModalRight';
+import ButtonCategoryCommon from '@/shared/components/ButtonCategoryCommon';
+import CategoryMsetUrlInfo from '@/shared/components/CategoryMsetUrlInfo';
+import CategoryTabSelect from '@/shared/components/CategoryTabSelect';
+import DropdownCategory from '@/shared/components/DropdownCategory';
 
 import { getTabName } from '@/shared/apis/modal/axios';
+import { useCategoryLists, useGetMsets } from '@/shared/apis/modal/queries';
+
+import { CATEGORY_MODALTABS } from '@/shared/constants/tabSelections';
+
+import AddBtn from '@/shared/assets/svgs/add_btn.svg?react';
+import MinusBtn from '@/shared/assets/svgs/minus_btn.svg?react';
+
+import CategoryCommonMoribSet from './CategoryCommonMoribSet';
+import InputCategoryUrl from './InputCategoryUrl';
 
 interface UrlInfo {
 	url: string;
@@ -35,6 +46,24 @@ const AddCategoryListModal = ({
 }: CategoryListModalProp) => {
 	const [isClicked, setIsClicked] = useState(false);
 	const [selectedOption, setSelectedOption] = useState('카테고리 추가');
+	const [selectedTabId, setSelectedTabId] = useState(CATEGORY_MODALTABS[0].id);
+
+	const { data: categoryData } = useCategoryLists();
+	const categories = categoryData?.data || [];
+	const [categoryId, setCategoryId] = useState<number>(0);
+
+	const { data: msetsList } = useGetMsets(categoryId);
+	const msetUrlInfos = msetsList || [];
+
+	const handleOptionId = (id: number) => {
+		setCategoryId(id);
+	};
+
+	const handleTabChange = (tab: number) => {
+		setSelectedTabId(tab);
+	};
+
+	const disabled = selectedTabId === 2;
 
 	const handleUrlInputChange = async (url: string) => {
 		try {
@@ -65,22 +94,105 @@ const AddCategoryListModal = ({
 	return (
 		<dialog ref={dialogRef} className="rounded-[10px]">
 			<div className="flex">
-				<CategoryModalLeft
-					handleClickButton={handleClickButton}
-					handleSelectOption={handleSelectOption}
-					isClicked={isClicked}
-					selectedOption={selectedOption}
-					handleRightModalUrlInfos={handleRightModalUrlInfos}
-				/>
-				<CategoryModalRight
-					rightModalUrlInfos={rightModalUrlInfos}
-					handleUrlInputChange={(url: string) => handleUrlInputChange(url)}
-					handleDeleteUrlInfo={(url: UrlInfo) => handleDeleteUrlInfo(url)}
-					handleClose={handleClose}
-					handleSubmitModal={handleSubmitModal}
-					moribSetName={moribSetName}
-					handleClearModalData={handleClearModalData}
-				/>
+				<div className="h-[80rem] w-[68.8rem] rounded-l-[10px] bg-gray-bg-04 py-[2.8rem] pl-[4.4rem] pr-[4.3rem]">
+					<header className="mb-[3.3rem]">
+						<h1 className="head-bold-24 text-gray-04">카테고리 추가</h1>
+					</header>
+					<aside className="mb-[8px]">
+						<div className="my-[8px]">
+							<CategoryTabSelect
+								tabs={CATEGORY_MODALTABS}
+								handleTabChange={handleTabChange}
+								selectedTabId={selectedTabId}
+							/>
+						</div>
+
+						<div className="relative mt-[0px]">
+							<DropdownCategory
+								optionData={categories}
+								disabled={disabled}
+								handleOptionId={handleOptionId}
+								handleClickButton={handleClickButton}
+								handleSelectOption={handleSelectOption}
+								isClicked={isClicked}
+								selectedOption={selectedOption}
+							/>
+						</div>
+					</aside>
+
+					<CategoryCommonMoribSet variant="smallLeft" urlInfos={msetUrlInfos}>
+						{selectedTabId !== 2 &&
+							msetUrlInfos.map((urlInfo: UrlInfo) => (
+								<div
+									key={urlInfo.url}
+									className="group flex h-[4.6rem] w-[100%] gap-[1.2rem] border-b border-gray-bg-04 px-[0.8rem] hover:bg-gray-bg-04"
+									onClick={() => handleRightModalUrlInfos(urlInfo)}
+								>
+									<CategoryMsetUrlInfo urlInfo={urlInfo} variant="smallLeft">
+										<div className="p-[1.25rem]">
+											<button type="button">
+												<AddBtn className="fill-gray-bg-07 group-hover:fill-mint-02-hover group-active:fill-mint-02-press" />
+											</button>
+										</div>
+									</CategoryMsetUrlInfo>
+								</div>
+							))}
+					</CategoryCommonMoribSet>
+				</div>
+				<div className="flex h-[80rem] w-[61.2rem] flex-col items-end justify-between rounded-r-[1rem] bg-gray-bg-03 pb-[3rem] pl-[3rem] pr-[4.3rem] pt-[9.7rem]">
+					<header className="subhead-bold-22 mb-[8px] flex w-full flex-row justify-start p-[1rem]">
+						<h2 className="text-mint-01">
+							{moribSetName.length > 0 ? (
+								moribSetName
+							) : (
+								<>
+									<h2 className="pr-[1rem]" /> _______ <span className="pr-[0.5rem]" />
+								</>
+							)}
+						</h2>
+						<h2 className="text-white">의 모립세트</h2>
+					</header>
+
+					<InputCategoryUrl
+						rightModalUrlInfos={rightModalUrlInfos}
+						variant="small"
+						onUrlInputChange={(url: string) => handleUrlInputChange(url)}
+					/>
+					<CategoryCommonMoribSet urlInfos={rightModalUrlInfos} variant="smallRight">
+						{rightModalUrlInfos.map((urlInfo, url) => (
+							<div key={url} className="flex h-[4.6rem] gap-[2rem] border-b border-gray-bg-04 px-[0.8rem]">
+								<CategoryMsetUrlInfo urlInfo={urlInfo} variant="smallRight">
+									<div className="p-[1.25rem]">
+										<button type="button" onClick={() => handleDeleteUrlInfo(urlInfo)}>
+											<MinusBtn className="fill-gray-bg-07 hover:fill-error-01 active:fill-error-03" />
+										</button>
+									</div>
+								</CategoryMsetUrlInfo>
+							</div>
+						))}
+					</CategoryCommonMoribSet>
+
+					<div className="mt-[3rem] flex gap-[16px]">
+						<ButtonCategoryCommon
+							variant="취소"
+							onClick={() => {
+								handleClearModalData();
+								handleClose();
+							}}
+						>
+							취소
+						</ButtonCategoryCommon>
+						<ButtonCategoryCommon
+							variant="완료"
+							onClick={() => {
+								handleClearModalData();
+								handleSubmitModal();
+							}}
+						>
+							완료
+						</ButtonCategoryCommon>
+					</div>
+				</div>
 			</div>
 		</dialog>
 	);
