+++ src/shared/components/ButtonCategoryCommon.tsx
@@ -3,20 +3,9 @@ import { ButtonHTMLAttributes, ReactNode } from 'react';
 interface CategoryBtnProps extends ButtonHTMLAttributes<HTMLButtonElement> {
 	variant?: '취소' | '완료';
 	children: ReactNode;
-	handleCloseModal?: () => void;
-	handleSubmit?: () => void;
-	handleClearModalData?: () => void;
 }
 
-const ButtonCategoryCommon = ({
-	disabled,
-	variant,
-	children,
-	handleCloseModal,
-	handleSubmit,
-	handleClearModalData,
-	...props
-}: CategoryBtnProps) => {
+const ButtonCategoryCommon = ({ disabled, variant, children, onClick, ...props }: CategoryBtnProps) => {
 	const btnVariant = {
 		취소: 'text-white bg-gray-bg-06 hover:bg-gray-bg-04 active:bg-gray-bg-05',
 		완료: 'text-gray-bg-01 bg-mint-02 hover:bg-mint-02-hover active:bg-mint-02-press',
@@ -27,23 +16,8 @@ const ButtonCategoryCommon = ({
 
 	const styledBtn = disabled ? disabledBtn : variant ? btnVariant[variant] : '';
 
-	const handleClick = () => {
-		if (handleSubmit) {
-			handleSubmit();
-		} else if (handleCloseModal) {
-			handleCloseModal();
-		}
-	};
 	return (
-		<button
-			className={`${styledBtn} ${commonStyle}`}
-			disabled={disabled}
-			{...props}
-			onClick={() => {
-				handleClick();
-				handleClearModalData?.();
-			}}
-		>
+		<button className={`${styledBtn} ${commonStyle}`} disabled={disabled} {...props} onClick={onClick}>
 			{children}
 		</button>
 	);
